Simplified the subclass behavior for Project.
The Project was using config.SERVER_URL for creating clients and that
was trouble for subclasses which may want to use a diffent server url.
Fixed the issue by adding a SERVER_URL attribute to Project class and
always accessing it from the class instead of from config. This allows
the subclasses to provide their own SERVER_URL in the sub class.